backend/requirements.txt
pytorch>=1.5.0
torchvision
pillow
